whisper_speech_to_text/requirements.txt
openai-whisper
librosa